Refactor doctor and nurse layouts to match patient panel
